emir-update: remove special-case for built-in nxml
Like for non-built-in packages that lack a library by the same name,
"nxml"'s `library' slot is now set explicitly.  Unlike mirrored
packages built-in packages (except the pseudo-package "emacs") always
have that slots `library' and `libraries', set automatically.  This
happens in `emir-update' (not in `emir-add') because new libraries may
be added and existing libraries may be moved or removed.

Now `emir-update' tries to automatically determine the new value for
`library' by locking for a library with a matching name and only if
that fails, then it used the old value.  In the case of "nxml" the
first step fails and the old explicitly set value is used.

If a library "nxml.el" appeared, then this would automatically be
taken into account.  On the other hand, if "nxml-mode.el" were moved,
then that would result in an error later in `epkg-update'.  But that
was already the case before.