Remove emir-migrate-github-package
To update all data the package has to be fully remove and re-added.
It would be possible to update everything very carefully without
actually removing the package.  But doing it right would be a lot of
work, and the only benefit would be that we would not occasionally
have to manually restore some manually created additional information.