emir-add: set repopage and homepage
Now all slots that can be set based solely on the `url' are being set
here.  Previously these two slots were only set in `emir-update'.
While `repopage' no longer is set there, `homepage' still is because
we might be able to extract a better value extracted from the main
library.  `emir--homepage' continues to fall back to the `url'-based
value because a maintainer may *remove* a previously explicit set
value, and in that case we have to reset to the default, because s/he
might have *deleted* the old homepage.