emir-clone: consolidate methods
Use the same method for all classes except `epkg-file-package'.

We cannot use the new generic method for that class too because there
is nothing to be cloned in that case.  Instead we have to initialize a
new repository.

We can use the same method for all other classes, by first cloning the
repository using `git clone' and only then calling `git submodule add'.
Previously we cloned using `git submodule add' for some classes, but
that's not possible for all classes.