emir-gh-init: handle mirroring github forks
Sometimes a new package is bootstrapped by forking another but has
never-the-less involved into another package in its own right, and
both packages should be mirrored.

Github does not allow a user or organization to contain more than one
fork of any repository in the same "fork-tree".  So if we mirror more
than one package from the same fork-tree, all but one of the mirrors
cannot be forks.

Instead of forking, we have to create a new repository with no
fork-relationship.  This commit automates that.  Previously we got an
error, then had to investigate, remove the failed import, comment out
some code, re-import, and restore the full code.  That didn't happen
very often, but often enough to by annoying.