emir--main-library: relax for shelved packages
A library is supposed to provide the matching feature, but for many
packages in the attic this is not the case, in fact in many cases that
is the reason those packages are in the attic.  We still have to be
able to figure out under what licenses these packages are released, so
don't insist on that proper feature being provided for shelved
packages.