Support mirroring epkg-elpa-core-package packages
At the time `emir-add-gelpa-packages' is run, those packages are
already being tracked as `emir-builtin-package' packages.  Delete
the respective information from the database and instead mirror
these packages semi-properly.

Unfortunately `epkg-elpa-core-package' currently derives from
`epkg-minority-package', which derives from `epkg-file-package',
which:
(1) Only supports a single file.
(2) Does not import the history of the file (using `git filter-branch'
    but instead imports a snapshot when the package is updated.

`soap-client' cannot be tracked as an `epkg-elpa-core-package'
because it consists of more than one file, and it is to be expected
that more such packages will be added.

Eventually `epkg-elpa-core-package' should work similarly to
`epkg-wiki-package', but that is a huge amount of work.

Also see https://github.com/emacscollective/borg/issues/85.