emir-clone: Fix epkg-gnu-elpa-package cloning regression
We pull from a local copy of the GNU Elpa repository to speed
up things.  In that repository we don't create local branches
for all the package branches so we have to pull from
"refs/remotes/origin/external/<package>".

"git clone --branch" expects a name, not a full ref, so have
to clone from the remote upstream repository.  We then adjust
the url so that subsequent pulls use the local clone instead.