emir-setup: Handle subrepo packages
Or rather consider such a package to be setup without creating the
"source" repository and delaying its creation until pulling happens.

This makes sense because others might want to checkout the submodules
without having to suffer through the pain that is "git subrepo", which
is only needed if one actually maintains a mirror.