emir-gh-init: Demote certain non-fatal errors
These heisenbugs occurred a lot.  They could be causes by a defect
of my own code, in url.el or on the server side.  My guess is that
it is a combination of the latter two.  The symptom is that our code
receives an empty response buffer.  The reason for this might be that
Github sends a weird response, which url.el cannot handle properly.